python/requirements.txt
uvicorn==0.30.5
fastapi==0.112.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
ollama==0.3.1
